# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-admin-api 后台管理接口文档",
        "title": "gin-admin-api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/dept": {
            "post": {
                "description": "创建部门；部门层级 fullId/fullName 由后端根据 parentId 自动维护",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "创建部门",
                "parameters": [
                    {
                        "description": "创建部门参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_dept_dto.CreateSysDeptDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/dept/detail/{id}": {
            "get": {
                "description": "根据部门 ID 获取部门详情，自动应用数据权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "获取部门详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "部门详情",
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_dept_vo.SysDeptVO"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/dept/list": {
            "get": {
                "description": "根据部门名称和状态获取部门列表，自动应用数据权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "获取部门列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态：1正常，2禁用，0全部",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "部门列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gin-admin-api_internal_api_dept_vo.SysDeptVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/dept/page": {
            "post": {
                "description": "根据部门名称、状态、上级部门分页查询部门列表，自动应用数据权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "分页获取部门",
                "parameters": [
                    {
                        "description": "部门分页查询参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysDeptPageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "部门分页列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gin-admin-api_internal_api_dept_vo.SysDeptVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/dept/{id}": {
            "put": {
                "description": "根据部门 ID 修改部门基础信息和父级关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "修改部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改部门参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ModifySysDeptDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据部门 ID 删除部门；存在子部门或账号时不能删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门中心"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ModifySysDeptDTO": {
            "type": "object",
            "required": [
                "fullId",
                "fullName",
                "id",
                "name",
                "parentId"
            ],
            "properties": {
                "createdBy": {
                    "description": "创建人",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 60
                },
                "fullId": {
                    "description": "全层级ID，例：1,5,12",
                    "type": "string",
                    "maxLength": 255
                },
                "fullName": {
                    "description": "全层级名称",
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "description": "主键id",
                    "type": "integer",
                    "minimum": 1
                },
                "leaderId": {
                    "description": "负责人账号id，关联sys_account.id",
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "description": "部门名称",
                    "type": "string",
                    "maxLength": 50
                },
                "parentId": {
                    "description": "上级部门id",
                    "type": "integer",
                    "minimum": 1
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string",
                    "maxLength": 20
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态1正常 2禁用",
                    "type": "integer"
                },
                "updatedBy": {
                    "description": "更新人",
                    "type": "integer"
                }
            }
        },
...
        "dto.SysDeptPageDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "部门名称，模糊查询",
                    "type": "string"
                },
                "pageNumber": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "parentId": {
                    "description": "上级部门id，默认0查询根部门",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用；0表示全部",
                    "type": "integer"
                }
            }
        },
...
        "gin-admin-api_internal_api_dept_dto.CreateSysDeptDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 60
                },
                "leaderId": {
                    "description": "负责人账号id，关联sys_account.id",
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "description": "部门名称",
                    "type": "string",
                    "maxLength": 50
                },
                "parentId": {
                    "description": "上级部门id，0表示根部门",
                    "type": "integer",
                    "minimum": 0
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string",
                    "maxLength": 20
                },
                "sort": {
                    "description": "排序，值越小越靠前",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用；不传默认1",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "gin-admin-api_internal_api_dept_vo.SysDeptVO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间，Unix秒",
                    "type": "integer"
                },
                "createdBy": {
                    "description": "创建人账号id",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "fullId": {
                    "description": "全层级部门ID，例：1,5,12",
                    "type": "string"
                },
                "fullName": {
                    "description": "全层级部门名称",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "leaderId": {
                    "description": "负责人账号id，关联sys_account.id",
                    "type": "integer"
                },
                "name": {
                    "description": "部门名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "上级部门id，0表示根部门",
                    "type": "integer"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "sort": {
                    "description": "排序，值越小越靠前",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间，Unix秒",
                    "type": "integer"
                },
                "updatedBy": {
                    "description": "更新人账号id",
                    "type": "integer"
                }
            }
        },
...
    }
}